Move components to named files
